Migrating functions to the executor.
