+++ crates/rooch-executor/Cargo.toml
@@ -32,14 +32,17 @@ tracing-subscriber = { workspace = true }
 schemars = { workspace = true }
 serde_with = { workspace = true }
 log = { workspace = true }
+itertools = { workspace = true }
 
 move-core-types = { workspace = true }
 move-resource-viewer = { workspace = true }
+move-binary-format = { workspace = true }
 
 moveos = { workspace = true }
 moveos-store = { workspace = true }
 moveos-types = { workspace = true }
 moveos-common = { workspace = true }
+moveos-verifier = { workspace = true }
 
 rooch-types = { workspace = true }
 rooch-framework = { workspace = true }
