+++ crates/rooch-executor/src/actor/executor.rs
@@ -14,9 +14,16 @@ use accumulator::inmemory::InMemoryAccumulator;
 use anyhow::Result;
 use async_trait::async_trait;
 use coerce::actor::{context::ActorContext, message::Handler, Actor};
+use itertools::Itertools;
+use move_binary_format::errors::{Location, PartialVMError, VMResult};
 use move_core_types::account_address::AccountAddress;
-use move_core_types::vm_status::VMStatus;
+use move_core_types::identifier::{IdentStr, Identifier};
+use move_core_types::language_storage::ModuleId;
+use move_core_types::resolver::ModuleResolver;
+use move_core_types::value::MoveValue;
+use move_core_types::vm_status::{StatusCode, VMStatus};
 use move_resource_viewer::MoveValueAnnotator;
+use moveos::gas::table::{initial_cost_schedule, MoveOSGasMeter};
 use moveos::moveos::{GasPaymentAccount, MoveOS};
 use moveos::vm::vm_status_explainer::explain_vm_status;
 use moveos_store::transaction_store::TransactionStore;
@@ -26,16 +33,17 @@ use moveos_types::function_return_value::AnnotatedFunctionReturnValue;
 use moveos_types::genesis_info::GenesisInfo;
 use moveos_types::h256::H256;
 use moveos_types::module_binding::MoveFunctionCaller;
-use moveos_types::move_types::as_struct_tag;
+use moveos_types::move_types::{as_struct_tag, FunctionId};
 use moveos_types::moveos_std::event::AnnotatedEvent;
 use moveos_types::moveos_std::event::EventHandle;
 use moveos_types::moveos_std::tx_context::TxContext;
 use moveos_types::state::{AnnotatedState, State};
 use moveos_types::state_resolver::{AnnotatedStateReader, StateReader};
-use moveos_types::transaction::TransactionExecutionInfo;
-use moveos_types::transaction::TransactionOutput;
 use moveos_types::transaction::VerifiedMoveOSTransaction;
 use moveos_types::transaction::{FunctionCall, MoveAction};
+use moveos_types::transaction::{MoveOSTransaction, TransactionExecutionInfo};
+use moveos_types::transaction::{TransactionOutput, VerifiedMoveAction};
+use moveos_verifier::metadata::load_module_metadata;
 use rooch_genesis::RoochGenesis;
 use rooch_store::RoochStore;
 use rooch_types::address::MultiChainAddress;
@@ -133,14 +141,14 @@ impl ExecutorActor {
 
         let vm_result = self.validate_authenticator(&moveos_tx.ctx, authenticator)?;
 
-        let can_pay_gas = self.moveos.execute_gas_validate(&moveos_tx)?;
+        let can_pay_gas = self.validate_gas_function(&moveos_tx)?;
 
         let mut pay_by_module_account = false;
         let mut gas_payment_account = moveos_tx.ctx.sender;
 
         if let Some(pay_gas) = can_pay_gas {
             if pay_gas {
-                let account_balance = self.moveos.get_address_balance(&moveos_tx)?;
+                let account_balance = self.get_account_balance(&moveos_tx)?;
                 let module_account = {
                     match &moveos_tx.action {
                         MoveAction::Function(call) => Some(*call.function_id.module_id.address()),
@@ -250,6 +258,171 @@ impl ExecutorActor {
             Err(vm_status) => Err(vm_status),
         };
         Ok(vm_result)
+    }
+
+    pub fn validate_gas_function(&self, tx: &MoveOSTransaction) -> VMResult<Option<bool>> {
+        let MoveOSTransaction { ctx, .. } = tx;
+
+        let cost_table = initial_cost_schedule();
+        let mut gas_meter = MoveOSGasMeter::new(cost_table, ctx.max_gas_amount);
+        gas_meter.set_metering(false);
+
+        let verified_moveos_action = self.moveos().verify(tx.clone())?;
+        let verified_action = verified_moveos_action.action;
+
+        match verified_action {
+            VerifiedMoveAction::Function { call } => {
+                let module_id = &call.function_id.module_id;
+                let loaded_module_bytes_result =
+                    self.moveos().moveos_resolver().get_module(module_id);
+                let loaded_module_bytes = match loaded_module_bytes_result {
+                    Ok(loaded_module_bytes_opt) => match loaded_module_bytes_opt {
+                        None => {
+                            return Err(PartialVMError::new(StatusCode::RESOURCE_DOES_NOT_EXIST)
+                                .with_message(
+                                    "The name of the gas_validate_function does not exist."
+                                        .to_string(),
+                                )
+                                .finish(Location::Module(module_id.clone())));
+                        }
+                        Some(module_bytes) => module_bytes,
+                    },
+                    Err(error) => {
+                        return Err(PartialVMError::new(StatusCode::RESOURCE_DOES_NOT_EXIST)
+                            .with_message(format!(
+                                "Load module data from module_id {:} was failed {:}.",
+                                module_id.clone(),
+                                error
+                            ))
+                            .finish(Location::Module(module_id.clone())));
+                    }
+                };
+
+                let module_metadata = load_module_metadata(module_id, Ok(loaded_module_bytes))?;
+                let gas_free_function_info = {
+                    match module_metadata {
+                        None => None,
+                        Some(runtime_metadata) => Some(runtime_metadata.gas_free_function_map),
+                    }
+                };
+
+                let called_function_name = call.function_id.function_name.to_string();
+                match gas_free_function_info {
+                    None => Ok(None),
+                    Some(gas_func_info) => {
+                        let full_called_function = format!(
+                            "0x{}::{}::{}",
+                            call.function_id.module_id.address().to_hex(),
+                            call.function_id.module_id.name(),
+                            called_function_name
+                        );
+                        let gas_func_info_opt = gas_func_info.get(&full_called_function);
+
+                        if let Some(gas_func_info) = gas_func_info_opt {
+                            let gas_validate_func_name = gas_func_info.gas_validate.clone();
+
+                            let split_function = gas_validate_func_name.split("::").collect_vec();
+                            if split_function.len() != 3 {
+                                return Err(PartialVMError::new(StatusCode::VM_EXTENSION_ERROR)
+                                    .with_message(
+                                        "The name of the gas_validate_function is incorrect."
+                                            .to_string(),
+                                    )
+                                    .finish(Location::Module(call.clone().function_id.module_id)));
+                            }
+                            let real_gas_validate_func_name =
+                                split_function.get(2).unwrap().to_string();
+
+                            let gas_validate_func_call = FunctionCall::new(
+                                FunctionId::new(
+                                    call.function_id.module_id.clone(),
+                                    Identifier::new(real_gas_validate_func_name).unwrap(),
+                                ),
+                                vec![],
+                                vec![],
+                            );
+
+                            let function_execution_result =
+                                self.moveos().execute_view_function(gas_validate_func_call);
+
+                            return if function_execution_result.vm_status == VMStatus::Executed {
+                                let return_value = function_execution_result.return_values.unwrap();
+                                if !return_value.is_empty() {
+                                    let first_return_value = return_value.get(0).unwrap();
+                                    Ok(Some(
+                                        bcs::from_bytes::<bool>(first_return_value.value.as_slice())
+                                            .expect(
+                                                "the return value of gas validate function should be bool",
+                                            ),
+                                    ))
+                                } else {
+                                    return Err(PartialVMError::new(
+                                        StatusCode::VM_EXTENSION_ERROR,
+                                    )
+                                    .with_message(
+                                        "the return value of gas_validate_function is empty."
+                                            .to_string(),
+                                    )
+                                    .finish(Location::Module(call.clone().function_id.module_id)));
+                                }
+                            } else {
+                                Ok(None)
+                            };
+                        };
+
+                        Ok(None)
+                    }
+                }
+            }
+            _ => Ok(None),
+        }
+    }
+
+    pub fn get_account_balance(&self, tx: &MoveOSTransaction) -> VMResult<u128> {
+        let MoveOSTransaction { ctx, .. } = tx;
+
+        let cost_table = initial_cost_schedule();
+        let mut gas_meter = MoveOSGasMeter::new(cost_table, ctx.max_gas_amount);
+        gas_meter.set_metering(false);
+
+        let verified_moveos_action = self.moveos().verify(tx.clone())?;
+        let verified_action = verified_moveos_action.action;
+
+        match verified_action {
+            VerifiedMoveAction::Function { call } => {
+                let module_address = call.function_id.module_id.address();
+
+                let gas_coin_module_id = ModuleId::new(
+                    AccountAddress::from_hex_literal("0x3").unwrap(),
+                    Identifier::from(IdentStr::new("gas_coin").unwrap()),
+                );
+                let gas_balance_func_call = FunctionCall::new(
+                    FunctionId::new(gas_coin_module_id, Identifier::new("balance").unwrap()),
+                    vec![],
+                    vec![MoveValue::Address(*module_address)
+                        .simple_serialize()
+                        .unwrap()],
+                );
+
+                let function_execution_result =
+                    self.moveos().execute_view_function(gas_balance_func_call);
+
+                if function_execution_result.vm_status == VMStatus::Executed {
+                    let return_value = function_execution_result.return_values.unwrap();
+                    let first_return_value = return_value.get(0).unwrap();
+
+                    let balance = bcs::from_bytes::<move_core_types::u256::U256>(
+                        first_return_value.value.as_slice(),
+                    )
+                    .expect("the return value of gas validate function should be u128");
+
+                    Ok(balance.unchecked_as_u128())
+                } else {
+                    Ok(0)
+                }
+            }
+            _ => Ok(0),
+        }
     }
 
     pub fn get_rooch_store(&self) -> RoochStore {
