+++ moveos/moveos/src/moveos.rs
@@ -9,7 +9,6 @@ use itertools::Itertools;
 use move_binary_format::errors::VMError;
 use move_binary_format::errors::{vm_status_of_result, Location, PartialVMError, VMResult};
 use move_core_types::identifier::IdentStr;
-use move_core_types::language_storage::ModuleId;
 use move_core_types::value::MoveValue;
 use move_core_types::vm_status::{KeptVMStatus, VMStatus};
 use move_core_types::{
@@ -265,152 +264,6 @@ impl MoveOS {
                     self.execution_cleanup(session, vm_err.into_vm_status(), None)
                 }
             }
-        }
-    }
-
-    pub fn get_address_balance(&self, tx: &MoveOSTransaction) -> VMResult<u128> {
-        let MoveOSTransaction {
-            ctx,
-            action,
-            pre_execute_functions: _pre_execute_functions,
-            post_execute_functions: _post_execute_functions,
-        } = tx;
-
-        let cost_table = initial_cost_schedule();
-        let mut gas_meter = MoveOSGasMeter::new(cost_table, ctx.max_gas_amount);
-        gas_meter.set_metering(false);
-
-        let mut session = self
-            .vm
-            .new_readonly_session(&self.db, ctx.clone(), gas_meter);
-
-        let verified_action = session.verify_move_action(action.clone())?;
-
-        match verified_action {
-            VerifiedMoveAction::Function { call } => {
-                let module_address = call.function_id.module_id.address();
-
-                let gas_coin_module_id = ModuleId::new(
-                    AccountAddress::from_hex_literal("0x3").unwrap(),
-                    Identifier::from(IdentStr::new("gas_coin").unwrap()),
-                );
-                let gas_validate_func_call = FunctionCall::new(
-                    FunctionId::new(gas_coin_module_id, Identifier::new("balance").unwrap()),
-                    vec![],
-                    vec![MoveValue::Address(*module_address)
-                        .simple_serialize()
-                        .unwrap()],
-                );
-
-                let return_value =
-                    session.execute_function_bypass_visibility(gas_validate_func_call)?;
-
-                let first_return_value = return_value.get(0).unwrap();
-
-                let balance = bcs::from_bytes::<move_core_types::u256::U256>(
-                    first_return_value.value.as_slice(),
-                )
-                .expect("the return value of gas validate function should be u128");
-
-                Ok(balance.unchecked_as_u128())
-            }
-            _ => Ok(0),
-        }
-    }
-
-    pub fn execute_gas_validate(&self, tx: &MoveOSTransaction) -> VMResult<Option<bool>> {
-        let MoveOSTransaction {
-            ctx,
-            action,
-            pre_execute_functions: _pre_execute_functions,
-            post_execute_functions: _post_execute_functions,
-        } = tx;
-
-        let cost_table = initial_cost_schedule();
-        let mut gas_meter = MoveOSGasMeter::new(cost_table, ctx.max_gas_amount);
-        gas_meter.set_metering(false);
-
-        let mut session = self
-            .vm
-            .new_readonly_session(&self.db, ctx.clone(), gas_meter);
-
-        let verified_action = session.verify_move_action(action.clone())?;
-
-        match verified_action {
-            VerifiedMoveAction::Function { call } => {
-                let module_id = &call.function_id.module_id;
-                let loaded_module_bytes = session.get_data_store().load_module(module_id);
-
-                let module_metadata = load_module_metadata(module_id, loaded_module_bytes)?;
-                let gas_free_function_info = {
-                    match module_metadata {
-                        None => None,
-                        Some(runtime_metadata) => Some(runtime_metadata.gas_free_function_map),
-                    }
-                };
-
-                let called_function_name = call.function_id.function_name.to_string();
-                match gas_free_function_info {
-                    None => Ok(None),
-                    Some(gas_func_info) => {
-                        let full_called_function = format!(
-                            "0x{}::{}::{}",
-                            call.function_id.module_id.address().to_hex(),
-                            call.function_id.module_id.name(),
-                            called_function_name
-                        );
-                        let gas_func_info_opt = gas_func_info.get(&full_called_function);
-
-                        if let Some(gas_func_info) = gas_func_info_opt {
-                            let gas_validate_func_name = gas_func_info.gas_validate.clone();
-
-                            let split_function = gas_validate_func_name.split("::").collect_vec();
-                            if split_function.len() != 3 {
-                                return Err(PartialVMError::new(StatusCode::VM_EXTENSION_ERROR)
-                                    .with_message(
-                                        "The name of the gas_validate_function is incorrect."
-                                            .to_string(),
-                                    )
-                                    .finish(Location::Module(call.clone().function_id.module_id)));
-                            }
-                            let real_gas_validate_func_name =
-                                split_function.get(2).unwrap().to_string();
-
-                            let gas_validate_func_call = FunctionCall::new(
-                                FunctionId::new(
-                                    call.function_id.module_id.clone(),
-                                    Identifier::new(real_gas_validate_func_name).unwrap(),
-                                ),
-                                vec![],
-                                vec![],
-                            );
-
-                            let return_value = session
-                                .execute_function_bypass_visibility(gas_validate_func_call)?;
-
-                            return if !return_value.is_empty() {
-                                let first_return_value = return_value.get(0).unwrap();
-                                Ok(Some(
-                                    bcs::from_bytes::<bool>(first_return_value.value.as_slice())
-                                        .expect(
-                                        "the return value of gas validate function should be bool",
-                                    ),
-                                ))
-                            } else {
-                                return Err(PartialVMError::new(StatusCode::VM_EXTENSION_ERROR)
-                                    .with_message(
-                                        "the return value of gas_validate_function is empty."
-                                            .to_string(),
-                                    )
-                                    .finish(Location::Module(call.clone().function_id.module_id)));
-                            };
-                        }
-
-                        Ok(None)
-                    }
-                }
-            }
-            _ => Ok(None),
         }
     }
